master: format1-solution.py Initial Commit
Updated python file to successfully exploit format1.  The comments of this python file include notes on how to disable ASLR (remember to go back and RErun stack5) and the manual solution I had to grind towards.  This code is nasty, repetitive, nasty, AND repetitive... but it works!  It also works.

Changes to be committed:
	new file:   format1-solution.py